Rename the RLB 'metrics' module to 'private'
