+++ glean-core/rlb/examples/prototype.rs
@@ -7,7 +7,7 @@ use std::env;
 use once_cell::sync::Lazy;
 use tempfile::Builder;
 
-use glean::{metrics::PingType, ClientInfoMetrics, Configuration, Error};
+use glean::{private::PingType, ClientInfoMetrics, Configuration, Error};
 
 #[allow(non_upper_case_globals)]
 pub static PrototypePing: Lazy<PingType> =
